backend/requirements.txt
django==4.2.0
djangorestframework==3.14.0
social-auth-app-django==5.4.0
requests==2.31.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
markdown==3.5.1
PyGithub==2.1.1
scikit-learn==1.3.2
gunicorn==21.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-django==4.7.0
